Add files via upload
In this modified program, you can add information about the method of communication or advertisement when adding an item related to wheat or wood. The "communication" field is part of each category (wheat and wood) to represent how that specific variety or type is communicated or advertised. You can further extend and customize this program to capture more details about the adoption of advertisement and communication.